Move 10 re-usable functions from ML Classifier Comparison to ML Classifier Comparison functions only
The aim of the project is to create a program which takes any classification datasets (or 2 related datasets) and outputs summary of differences between classifiers trained from the datasets. Hence to make code independent from datasets, there must be a program file containing general functions only, which can be imported in main program.